Feat: 필터링 탭 수정, framer-motion 라이브러리 설치, 조금 구현 #5
