+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "env-cmd": "^10.1.0",
+    "framer-motion": "^11.9.0",
     "next": "14.2.5",
     "react": "^18",
     "react-dom": "^18",
