+++ src/app/(main)/works/[id]/page.js
@@ -1,3 +1,10 @@
+"use client";
+import { useParams } from "next/navigation";
+import styles from "./page.module.css";
+//
+//
+//
 export default function WorkDetailPage() {
-  return <main></main>;
+  const pathname = useParams();
+  return <main className={styles.main}>{pathname.id}</main>;
 }
